Moved create_lookup() out of utils and into disease_correlation
