+++ tests/test_utils.py
@@ -411,25 +411,6 @@ class TestCheckRequiredDatasetsAndColumns:
             utils.check_required_datasets_and_columns(datasets, self.required_input)
 
 
-class TestCreateLookup:
-    def test_create_lookup(self):
-        input_dataframe = pd.DataFrame(
-            [
-                {"A": "a1", "B": "b1", "C": "c1"},
-                {"A": "a1", "B": "b2", "C": "c1"},
-                {"A": "a1", "B": "b3", "C": "c1"},
-                {"A": "a2", "B": "b4", "C": "c2"},
-            ]
-        )
-        group_by_col = "A"
-        expected_output = {
-            "a1": {"B": ["b1", "b2", "b3"], "C": "c1"},
-            "a2": {"B": "b4", "C": "c2"},
-        }
-        output = utils.create_lookup(df=input_dataframe, group_by_col=group_by_col)
-        assert output == expected_output
-
-
 class TestFlattenList:
     def test_flatten_list_empty(self):
         assert utils.flatten_list([]) == []
